qa/functional/selenium/requirements.txt
allure-pytest==2.13.4
pytest==8.3.3
pytest-xdist==3.6.1
selenium==4.23.1
python-dotenv==1.0.1
